[Refactor] Curriculum 도메인 내 파일 구조 개선 (#723)
* refactor: CurriculumQueryService 에서 변수 이름 refactor

* refactor: 안쓰는 http로 API 테스트 기능 제거

* refactor: 안쓰는 API 제거 WorkbookController -> ChallengerWorkbookController로 이름 변경

* refactor: Controller이름 통일

* refactor: Swagger 설명 수정

* feat: 챌린저가 워크북 제출물 수정 하는 API Controller 추가

* refactor: 안쓰는 manage 메서드 제거

* refactor: command 패키지에 있는 command 파일들 dto 폴더 안으로 이동

* refactor: ManageWorkbookUseCase에 운영진구역, 챌린저 구역나누는 주석 추가, release 메서드 추가, 안쓰는 usecase제거

* refactor: release메서드 이동, ReleaseWorkbookUseCaseTest 제거

* refactor: ManageWorkbookUsecase -> ManageChallengerWorkbookUseCase로 이름 변경

* refactor: SubmitWorkbookCommand Deprecated, ManageChallengerWorkbookUseCase에 주석 설명 추가

* refactor: curriculum Query UseCase들 엔티티 단위(자원 단위)로 묶기

* refactor: 도메인에서 자원별로 service 분리
