--- http/curriculum/admin-curriculum.http
@@ -1,141 +1,0 @@
-###############################################################################
-# Admin Curriculum API - 목데이터 입력용
-#
-# 커리큘럼 관리 API (PUT)
-# - 모든 API는 자동으로 활성 기수(isActive=true)를 기준으로 동작
-# - 배포 환경에서 초기 데이터 세팅용으로 사용
-#
-# [주의] prod 환경 실행 전 반드시 확인!
-###############################################################################
-
-### [임시] 커리큘럼 생성 - 웹팀용 부분 데이터
-# 웹팀에서 사용하는 간소화된 커리큘럼 생성 요청입니다.
-# 화면에 있는 필드(weekNo, title, description)만 보내면 됩니다.
-#
-# [선택 필드 동작 방식]
-# - 신규 생성 (id 없음): startDate, endDate, missionType에 기본값 적용 (2099-12-31, LINK)
-# - 수정 (id 있음): 보내지 않은 필드는 기존값 유지
-#
-PUT {{host}}/api/v1/admin/curriculums
-Content-Type: application/json
-
-{
-  "part": "SPRINGBOOT",
-  "title": "9기 Springboot",
-  "workbooks": [
-    {
-      "weekNo": 1,
-      "title": "1주차 - Spring Boot 시작하기",
-      "description": "스프링 부트 기본 개념과 프로젝트 설정"
-    },
-    {
-      "weekNo": 2,
-      "title": "2주차 - Spring MVC",
-      "description": "컨트롤러, 서비스, 레포지토리 계층 이해"
-    },
-    {
-      "weekNo": 3,
-      "title": "3주차 - JPA 기초",
-      "description": "엔티티 설계와 CRUD 구현"
-    }
-  ]
-}
-
-### [임시] 워크북 수정 - 웹팀용 부분 데이터
-# id가 있는 워크북을 수정할 때, 화면에 있는 필드만 보내도 됩니다.
-# startDate, endDate, missionType 등 보내지 않은 필드는 기존값이 유지됩니다.
-PUT {{host}}/api/v1/admin/curriculums
-Content-Type: application/json
-
-{
-  "part": "SPRINGBOOT",
-  "title": "9기 Springboot",
-  "workbooks": [
-    {
-      "id": 1,
-      "weekNo": 1,
-      "title": "1주차 - 수정된 제목",
-      "description": "수정된 설명"
-    }
-  ]
-}
-
-### [정식] 커리큘럼 생성 - 전체 데이터
-# 모든 필드를 포함한 정식 커리큘럼 생성 요청입니다.
-# 앱에서 사용하려면 워크북 배포(release) API를 호출해야 합니다.
-PUT {{host}}/api/v1/admin/curriculums
-Content-Type: application/json
-
-{
-  "part": "SPRINGBOOT",
-  "title": "9기 Springboot",
-  "workbooks": [
-    {
-      "weekNo": 1,
-      "title": "1주차 - Spring Boot 시작하기",
-      "description": "스프링 부트 기본 개념과 프로젝트 설정",
-      "workbookUrl": "https://www.notion.so/workbook-1",
-      "startDate": "2024-03-01",
-      "endDate": "2024-03-07",
-      "missionType": "LINK"
-    },
-    {
-      "weekNo": 2,
-      "title": "2주차 - Spring MVC",
-      "description": "컨트롤러, 서비스, 레포지토리 계층 이해",
-      "workbookUrl": "https://www.notion.so/workbook-2",
-      "startDate": "2024-03-08",
-      "endDate": "2024-03-14",
-      "missionType": "LINK"
-    },
-    {
-      "weekNo": 3,
-      "title": "3주차 - JPA 기초",
-      "description": "엔티티 설계와 CRUD 구현",
-      "workbookUrl": "https://www.notion.so/workbook-3",
-      "startDate": "2024-03-15",
-      "endDate": "2024-03-21",
-      "missionType": "MEMO"
-    }
-  ]
-}
-
-### [정식] 기존 커리큘럼 수정
-# 기존 워크북의 id를 포함하여 수정합니다.
-# - id가 있는 워크북: 수정
-# - id가 없는 워크북: 신규 생성
-# - 요청에 없는 기존 워크북: 삭제 (단, 제출된 워크북은 삭제 불가)
-PUT {{host}}/api/v1/admin/curriculums
-Content-Type: application/json
-
-{
-  "part": "SPRINGBOOT",
-  "title": "9기 Springboot (수정됨)",
-  "workbooks": [
-    {
-      "id": 1,
-      "weekNo": 1,
-      "title": "1주차 - Spring Boot 시작하기 (수정)",
-      "description": "수정된 설명",
-      "workbookUrl": "https://www.notion.so/workbook-1-updated",
-      "startDate": "2024-03-01",
-      "endDate": "2024-03-07",
-      "missionType": "LINK"
-    },
-    {
-      "id": 2,
-      "weekNo": 2,
-      "title": "2주차 - Spring MVC",
-      "description": "컨트롤러, 서비스, 레포지토리 계층 이해"
-    },
-    {
-      "weekNo": 4,
-      "title": "4주차 - 새로 추가된 주차",
-      "description": "새로운 내용",
-      "workbookUrl": "https://www.notion.so/workbook-4",
-      "startDate": "2024-03-22",
-      "endDate": "2024-03-28",
-      "missionType": "PLAIN"
-    }
-  ]
-}
