--- http/http-client.env.json
@@ -1,11 +1,0 @@
-{
-  "local": {
-    "host": "http://localhost:8080"
-  },
-  "dev": {
-    "host": "https://dev-api.example.com"
-  },
-  "prod": {
-    "host": "https://api.example.com"
-  }
-}
