+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v1/swagger/AdminCurriculumQueryControllerApi.java
@@ -30,8 +30,8 @@ import java.util.List;
  *   <li>학교 파트장/기타 운영진: 본인 학교의 담당 파트만 조회 가능</li>
  * </ul>
  */
-@Tag(name = "Curriculum | 워크북 Query", description = "")
-public interface WorkbookQueryControllerApi {
+@Tag(name = "Curriculum | 운영진용 워크북 Query", description = "")
+public interface AdminCurriculumQueryControllerApi {
 
     @Operation(
         summary = "워크북 제출 현황 조회",
