+++ src/main/java/com/umc/product/curriculum/application/port/in/command/ManageChallengerWorkbookUseCase.java
@@ -1,12 +1,19 @@
 package com.umc.product.curriculum.application.port.in.command;
 
+import com.umc.product.curriculum.application.port.in.command.dto.ReviewWorkbookCommand;
+import com.umc.product.curriculum.application.port.in.command.dto.SelectBestWorkbookCommand;
+import com.umc.product.curriculum.application.port.in.command.dto.SubmitChallengerWorkbookCommand;
+import com.umc.product.curriculum.application.port.in.command.dto.SubmitWorkbookCommand;
+
 /**
  * 워크북 관리 UseCase
  * <p>
  * 챌린저의 워크북 제출 및 운영진의 워크북 검토/베스트 선정 기능을 제공합니다.
  */
-public interface ManageWorkbookUseCase {
+public interface ManageChallengerWorkbookUseCase {
 
+
+    // <------------------ 챌린저 전용 -------------------------->
     /**
      * 워크북 제출 (챌린저 전용)
      * <p>
@@ -29,10 +36,12 @@ public interface ManageWorkbookUseCase {
      */
     void submitByWorkbookId(SubmitChallengerWorkbookCommand command);
 
+
+    // <---------------------- 운영진 전용 -------------------------->
     /**
-     * 워크북 검토 (운영진 전용)
+     * 챌린저 워크북 검토 (운영진 전용)
      * <p>
-     * 운영진이 제출된 워크북을 검토하여 통과(PASS) 또는 반려(FAIL) 처리합니다.
+     * 운영진이 제출된 챌린저 워크북을 검토하여 통과(PASS) 또는 반려(FAIL) 처리합니다.
      *
      * @param command 검토 커맨드 (워크북 ID, 검토 결과, 피드백)
      */
