--- src/main/java/com/umc/product/curriculum/application/port/in/command/ManageCurriculumUseCase.java
@@ -1,6 +1,0 @@
-package com.umc.product.curriculum.application.port.in.command;
-
-public interface ManageCurriculumUseCase {
-
-    void manage(CurriculumCommand command);
-}
