--- src/main/java/com/umc/product/curriculum/application/port/in/command/ReleaseWorkbookUseCase.java
@@ -1,6 +1,0 @@
-package com.umc.product.curriculum.application.port.in.command;
-
-public interface ReleaseWorkbookUseCase {
-
-    void release(Long workbookId);
-}
