+++ src/main/java/com/umc/product/curriculum/application/port/in/command/dto/CurriculumCommand.java
@@ -1,4 +1,4 @@
-package com.umc.product.curriculum.application.port.in.command;
+package com.umc.product.curriculum.application.port.in.command.dto;
 
 import com.umc.product.common.domain.enums.ChallengerPart;
 import com.umc.product.curriculum.domain.enums.MissionType;
