+++ src/main/java/com/umc/product/curriculum/application/port/in/command/dto/SubmitWorkbookCommand.java
@@ -1,4 +1,4 @@
-package com.umc.product.curriculum.application.port.in.command;
+package com.umc.product.curriculum.application.port.in.command.dto;
 
 import jakarta.validation.constraints.NotNull;
 
@@ -9,6 +9,7 @@ import jakarta.validation.constraints.NotNull;
  * @param challengerId         제출하는 챌린저 ID
  * @param submission           제출 내용 (링크 또는 메모, PLAIN 타입인 경우 null 가능)
  */
+@Deprecated(since = "1.3.0", forRemoval = true)
 public record SubmitWorkbookCommand(
         @NotNull(message = "원본 워크북 ID는 필수입니다")
         Long originalWorkbookId,
