+++ src/test/java/com/umc/product/curriculum/application/port/in/command/ManageChallengerWorkbookUseCaseTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.umc.product.challenger.domain.Challenger;
 import com.umc.product.common.domain.enums.ChallengerPart;
+import com.umc.product.curriculum.application.port.in.command.dto.ReviewWorkbookCommand;
+import com.umc.product.curriculum.application.port.in.command.dto.SelectBestWorkbookCommand;
 import com.umc.product.curriculum.application.port.out.LoadChallengerWorkbookPort;
 import com.umc.product.curriculum.domain.ChallengerWorkbook;
 import com.umc.product.curriculum.domain.Curriculum;
@@ -25,10 +27,10 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
+class ManageChallengerWorkbookUseCaseTest extends UseCaseTestSupport {
 
     @Autowired
-    private ManageWorkbookUseCase manageWorkbookUseCase;
+    private ManageChallengerWorkbookUseCase manageChallengerWorkbookUseCase;
 
     @Autowired
     private GisuFixture gisuFixture;
@@ -164,7 +166,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.review(command);
+            manageChallengerWorkbookUseCase.review(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -185,7 +187,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.review(command);
+            manageChallengerWorkbookUseCase.review(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -205,7 +207,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.review(command);
+            manageChallengerWorkbookUseCase.review(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -218,7 +220,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             // given
             ChallengerWorkbook workbook = createWorkbookWithStatus(WorkbookStatus.SUBMITTED);
 
-            manageWorkbookUseCase.review(new ReviewWorkbookCommand(
+            manageChallengerWorkbookUseCase.review(new ReviewWorkbookCommand(
                 workbook.getId(),
                 WorkbookStatus.PASS,
                 null
@@ -231,7 +233,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
                 "다시 검토"
             );
 
-            assertThatThrownBy(() -> manageWorkbookUseCase.review(command))
+            assertThatThrownBy(() -> manageChallengerWorkbookUseCase.review(command))
                 .isInstanceOf(BusinessException.class);
         }
 
@@ -245,7 +247,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when & then
-            assertThatThrownBy(() -> manageWorkbookUseCase.review(command))
+            assertThatThrownBy(() -> manageChallengerWorkbookUseCase.review(command))
                 .isInstanceOf(BusinessException.class);
         }
     }
@@ -265,7 +267,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.selectBest(command);
+            manageChallengerWorkbookUseCase.selectBest(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -285,7 +287,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.selectBest(command);
+            manageChallengerWorkbookUseCase.selectBest(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -304,7 +306,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when
-            manageWorkbookUseCase.selectBest(command);
+            manageChallengerWorkbookUseCase.selectBest(command);
 
             // then
             ChallengerWorkbook result = loadChallengerWorkbookPort.findById(workbook.getId());
@@ -323,7 +325,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when & then
-            assertThatThrownBy(() -> manageWorkbookUseCase.selectBest(command))
+            assertThatThrownBy(() -> manageChallengerWorkbookUseCase.selectBest(command))
                 .isInstanceOf(BusinessException.class);
         }
 
@@ -336,7 +338,7 @@ class ManageWorkbookUseCaseTest extends UseCaseTestSupport {
             );
 
             // when & then
-            assertThatThrownBy(() -> manageWorkbookUseCase.selectBest(command))
+            assertThatThrownBy(() -> manageChallengerWorkbookUseCase.selectBest(command))
                 .isInstanceOf(BusinessException.class);
         }
     }
