--- src/test/java/com/umc/product/curriculum/application/port/in/command/ReleaseWorkbookUseCaseTest.java
@@ -1,103 +1,0 @@
-package com.umc.product.curriculum.application.port.in.command;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import com.umc.product.common.domain.enums.ChallengerPart;
-import com.umc.product.curriculum.application.port.out.LoadOriginalWorkbookPort;
-import com.umc.product.curriculum.application.port.out.SaveCurriculumPort;
-import com.umc.product.curriculum.domain.Curriculum;
-import com.umc.product.curriculum.domain.OriginalWorkbook;
-import com.umc.product.curriculum.domain.enums.MissionType;
-import com.umc.product.global.exception.BusinessException;
-import com.umc.product.organization.domain.Gisu;
-import com.umc.product.support.UseCaseTestSupport;
-import com.umc.product.support.fixture.GisuFixture;
-import java.time.Instant;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-@Disabled
-class ReleaseWorkbookUseCaseTest extends UseCaseTestSupport {
-
-    @Autowired
-    private ReleaseWorkbookUseCase releaseWorkbookUseCase;
-
-    @Autowired
-    private SaveCurriculumPort saveCurriculumPort;
-
-    @Autowired
-    private LoadOriginalWorkbookPort loadOriginalWorkbookPort;
-
-    @Autowired
-    private GisuFixture gisuFixture;
-
-    private Gisu activeGisu;
-
-    @BeforeEach
-    void setUp() {
-        activeGisu = gisuFixture.활성_기수(9L);
-    }
-
-    @Test
-    void 워크북을_배포한다() {
-        // given
-        Curriculum curriculum = Curriculum.create(activeGisu.getId(), ChallengerPart.SPRINGBOOT, "커리큘럼");
-        OriginalWorkbook workbook = OriginalWorkbook.create(
-            curriculum, 1, "1주차 - Spring 시작하기", "설명", "http://workbook.url",
-            Instant.parse("2024-03-01T00:00:00Z"), Instant.parse("2024-03-07T23:59:59Z"), MissionType.LINK
-        );
-        curriculum.addWorkbook(workbook);
-        Curriculum savedCurriculum = saveCurriculumPort.save(curriculum);
-        Long workbookId = savedCurriculum.getOriginalWorkbooks().get(0).getId();
-
-        // when
-        releaseWorkbookUseCase.release(workbookId);
-
-        // then
-        OriginalWorkbook releasedWorkbook = loadOriginalWorkbookPort.findById(workbookId);
-        assertThat(releasedWorkbook.isReleased()).isTrue();
-        assertThat(releasedWorkbook.getReleasedAt()).isNotNull();
-    }
-
-    @Test
-    void 존재하지_않는_워크북을_배포하면_예외가_발생한다() {
-        // given
-        Long nonExistentWorkbookId = 999L;
-
-        // when & then
-        assertThatThrownBy(() -> releaseWorkbookUseCase.release(nonExistentWorkbookId))
-            .isInstanceOf(BusinessException.class);
-    }
-
-    @Test
-    void 여러_워크북을_개별적으로_배포할_수_있다() {
-        // given
-        Curriculum curriculum = Curriculum.create(activeGisu.getId(), ChallengerPart.SPRINGBOOT, "커리큘럼");
-        OriginalWorkbook workbook1 = OriginalWorkbook.create(
-            curriculum, 1, "1주차", null, null,
-            Instant.parse("2024-03-01T00:00:00Z"), Instant.parse("2024-03-07T23:59:59Z"), MissionType.LINK
-        );
-        OriginalWorkbook workbook2 = OriginalWorkbook.create(
-            curriculum, 2, "2주차", null, null,
-            Instant.parse("2024-03-08T00:00:00Z"), Instant.parse("2024-03-14T23:59:59Z"), MissionType.LINK
-        );
-        curriculum.addWorkbook(workbook1);
-        curriculum.addWorkbook(workbook2);
-        Curriculum savedCurriculum = saveCurriculumPort.save(curriculum);
-        Long workbook1Id = savedCurriculum.getOriginalWorkbooks().get(0).getId();
-        Long workbook2Id = savedCurriculum.getOriginalWorkbooks().get(1).getId();
-
-        // when
-        releaseWorkbookUseCase.release(workbook1Id);
-
-        // then
-        OriginalWorkbook released1 = loadOriginalWorkbookPort.findById(workbook1Id);
-        OriginalWorkbook notReleased2 = loadOriginalWorkbookPort.findById(workbook2Id);
-
-        assertThat(released1.isReleased()).isTrue();
-        assertThat(notReleased2.isReleased()).isFalse();
-    }
-}
